+++ backend/src/app.js
@@ -32,6 +32,7 @@ const healthCheckRouter = require('./routes/healthCheck')
 const messageRouter = require('./routes/message')
 const notificationRouter = require('./routes/notification')
 const applicationsRouter = require('./routes/applications')
+const irregularApplicationsRouter = require('./routes/irregular')
 const organizationsRouter = require('./routes/organizations')
 const fundingAgreementsRouter = require('./routes/fundingAgreements')
 const facilitiesRouter = require('./routes/facilities')
@@ -234,6 +235,7 @@ app.use(morgan(config.get('server:morganFormat'), { stream: logStream }))
 app.use(/(\/api)?/, apiRouter)
 
 apiRouter.use('/applications', applicationsRouter)
+apiRouter.use('/irregular', irregularApplicationsRouter)
 apiRouter.use('/auth', authRouter)
 apiRouter.use('/config', configRouter)
 apiRouter.use('/documents', documentsRouter)
