+++ backend/src/components/applications.js
@@ -1,7 +1,7 @@
 'use strict'
 const { getOperation, patchOperationWithObjectId, postOperation, deleteOperationWithObjectId, handleError } = require('./utils')
 const { MappableObjectForFront, MappableObjectForBack } = require('../util/mapping/MappableObject')
-const { ApplicationMappings, ApplicationProviderEmployeeMappings, SupplementaryApplicationMappings, IrregularExpenseMappings } = require('../util/mapping/Mappings')
+const { ApplicationMappings, ApplicationProviderEmployeeMappings, SupplementaryApplicationMappings } = require('../util/mapping/Mappings')
 const { buildFilterQuery, buildDateFilterQuery } = require('../util/common')
 const HttpStatus = require('http-status-codes')
 const { isEmpty } = require('lodash')
@@ -161,23 +161,6 @@ async function createSupplementaryApplication(req, res) {
   }
 }
 
-async function getIrregularExpenseApplications(req, res) {
-  try {
-    console.log('called!!')
-    const applications = []
-    //const filterObj = { applicationId: req.params.applicationId, ...req?.query }
-    const operation = `ofm_expenses?$filter=(${buildFilterQuery(req?.query, IrregularExpenseMappings)})`
-    console.log(operation)
-    const response = await getOperation(operation)
-
-    response?.value?.forEach((application) => applications.push(new MappableObjectForFront(application, IrregularExpenseMappings).toJSON()))
-
-    return res.status(HttpStatus.OK).json(applications)
-  } catch (e) {
-    handleError(res, e)
-  }
-}
-
 async function updateSupplementaryApplication(req, res) {
   try {
     const payload = new MappableObjectForBack(req.body, SupplementaryApplicationMappings).toJSON()
@@ -209,29 +192,9 @@ async function deleteSupplementaryApplication(req, res) {
 
 async function getSupplementaryApplicationPDF(req, res) {
   try {
-    const operation = `ofm_allowances(${req.params.applicationId})/ofm_supplementaryapplicationpdf`
+    const operation = `ofm_allowances(${req.params.applicationId})/ofm_supplementary_application_pdf`
     const response = await getOperation(operation)
     return res.status(HttpStatus.OK).json(response?.value)
-  } catch (e) {
-    handleError(res, e)
-  }
-}
-
-async function getIrregularExpensePDF(req, res) {
-  try {
-    const operation = `ofm_expenses(${req.params.expenseApplicationId})/ofm_file`
-    const response = await getOperation(operation)
-    return res.status(HttpStatus.OK).json(response?.value)
-  } catch (e) {
-    handleError(res, e)
-  }
-}
-
-async function getIrregularExpenseByID(req, res) {
-  try {
-    const operation = `ofm_expenses(${req.params.expenseApplicationId})`
-    const response = await getOperation(operation)
-    return res.status(HttpStatus.OK).json(new MappableObjectForFront(response, IrregularExpenseMappings).toJSON())
   } catch (e) {
     handleError(res, e)
   }
@@ -298,9 +261,6 @@ module.exports = {
   deleteEmployeeCertificate,
   getApplicationPDF,
   getSupplementaryApplicationPDF,
-  getIrregularExpenseApplications,
   getSupplementaryApprovalPDF,
   getSupplementaryApplicationById,
-  getIrregularExpensePDF,
-  getIrregularExpenseByID,
 }
