+++ frontend/src/utils/constants.js
@@ -28,7 +28,7 @@ export const ApiRoutes = Object.freeze({
   REPORTS: baseRoot + '/reports',
   SUPPLEMENTARY_ALLOWANCES: baseRoot + '/supp-allowances',
   SUPPLEMENTARY_APPLICATIONS: baseRoot + '/applications/supplementary',
-  IRREGULAR_APPLICATIONS: baseRoot + '/applications/irregular',
+  IRREGULAR_APPLICATIONS: baseRoot + '/irregular',
   USER: baseRoot + '/user',
   USER_FACILITIES: baseRoot + '/user/:contactId/facilities',
   USER_PERMISSIONS_FACILITIES: baseRoot + '/user/permissions-facilities',
